[preview-server]: refactor webview
