+++ src/constants.ts
@@ -5,3 +5,4 @@ export const PLAYGROUND_FILE = "playground.json";
 export const UNTITLED_SCHEME = "untitled";
 export const URI_PATTERN = /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)/;
 export const ZERO_WIDTH_SPACE = "‎‎\u200B";
+export const CSS_ELEMENT_ID = `#gistpad-styles`;
