+++ src/interfaces/ITranspileFile.ts
@@ -1,0 +1,4 @@
+export interface ITranspileFile {
+  fileName: string;
+  content: string;
+}
